flight_management/requirements.txt
Django~=4.1.0
python-dotenv~=0.21.0
djangorestframework~=3.14.0
drf-spectacular~=0.25.0
psycopg2-binary~=2.9.0
drf-nested-routers~=0.93.0
django-filter~=22.0
